changed the types of parameters in the from_pixel method in GameObject in object.rs

+++ src/object.rs
@@ -54,13 +54,13 @@ pub struct GameObject {
 }
 
 impl GameObject {
-    pub fn from_pixel(x: f64, y: f64, scale_x: f32, scale_y: f32, color: [u8; 3], color_idx: usize, scale_multi: f32) -> Self {
+    pub fn from_pixel(x: u32, y: u32, scale_x: u32, scale_y: u32, color: [u8; 3], color_idx: usize, scale_multi: f32) -> Self {
         Self {
             id: 211,
-            x: (x + (scale_x - 1.) as f64 / 2.) * 30. * scale_multi as f64,
-            y: (0. - y - (scale_y - 1.) as f64 / 2.) * 30. * scale_multi as f64,
-            scale_x: scale_x * scale_multi,
-            scale_y: scale_y * scale_multi,
+            x: (x as f64 + (scale_x - 1) as f64 / 2.) * 30. * scale_multi as f64,
+            y: (0. - y as f64 - (scale_y - 1) as f64 / 2.) * 30. * scale_multi as f64,
+            scale_x: scale_x as f32 * scale_multi,
+            scale_y: scale_y as f32 * scale_multi,
             color_id: 1011,
             hsv: Hsv::rgb_to_hsv(color),
             layer: color_idx,
